requirements.txt
pyspark>=3.6
apache-airflow>=2.9.4